requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.11 
redis==5.0.1
celery==5.3.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
httpx==0.26.0
PyGithub==2.1.1
plotly==5.18.0
matplotlib==3.8.2
openpyxl==3.1.2
reportlab==4.0.9
jinja2==3.1.3
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0
